Feature

+++ modules/gdnative/include/videodecoder/godot_videodecoder.h
@@ -47,7 +47,7 @@ typedef struct
 	void *(*constructor)(godot_object *);
 	void (*destructor)(void *);
 	const char *(*get_plugin_name)(void);
-	const char **(*get_supported_extensions)(int *count);
+	const char **(*get_supported_extensions)(size_t *count);
 	godot_bool (*open_file)(void *, void *); // data struct, and a FileAccess pointer
 	godot_real (*get_length)(const void *);
 	godot_real (*get_playback_position)(const void *);
